refactor(config): LottoConfig 클래스명 변경
LottoConfig 를 LottoConfiguration 으로 변경했습니다.
